Implement basic color preferences
#2

Colors now evolve based on user selection. Entering a portal will change the colors of all portals. The selected portal changes to a random color. The exit portal selects a color at random that was not the selected portal color. All other portals are modified using Color.Lerp between the selected color and the current color of that portal.
Comments added to all code