Long Weekend Update
Work on issue #4
Debug methods added - these have helped / will help with tracking down problems
Some fairly long rewrites. Most more than once.
Portal evolution can now be rewound. There are still a few bugs in it -mostly with the display color of the exit portals not updating- but it does work and can return to the original room setup.
Portals are being transitioned to just contain methods to deal with the physical portal in the game. The logical 'portal' as a member of the population with a current expressed genetic trait (color) should be moved to Population. There is a task in VisStudio ready to start this process.

Work on issue #5
Portals have Sprite displaySprite now, but they are not implemented. This will eventually display the image created by the network.
Population/Portal/Genotype relationships have been restructured to make ready for this issue. Once a network is creating a sprite they can be displayed on the portal