HUD/Inventory - using saved items
#31

Items in the inventory can now be used to change the geno of a portal and redraw the image. This has highlighted an interesting "feature" showing that saved items are references. Overwritting a portal will also overwrite the saved geno and redraw that inventory slot thumbnail.

This may be a desireable feature:
saved portals can continue to evolve in the inventory bar - this could mean that using a portal is where the deep copy is made, so a used portal will disconnect from the evolution of the saved one, but the saved one will continue to evolve.
or, we could deep copy the geno on save and then they wont evolve on the bar