TWEANNGenotype deep copy
#55

This should be built, but now the mutations are acting very strange. I need to put markers in and track what is happening. On the first room change a perturb links looped through itself many times, not forever, just some arbitrary (possibly links.count times) number. Committing for now and will attempt to trace after the merge this weekend.