#83 Red bias in artworks
Red bias was from hue value coming back from the CPPN as an unbounded and signed value. This needs to be in the range of 0..1, but being unbounded means that the only solution was to truncate the value. Only values between 0..1 had any effect, everything else was converted to 0 or 1, both of which are red.

Solution (temp):
the entire cppn is processed and all values are stored in a secondary array. During that process, the values are compared against the running min and running max values that the CPPN has produced. Those are used to find the range of the CPPN (although this is NOT limited to the hue value, something we may change before testing). That range is used to remap the hue value from the cppn to 0...1 without destroying the values outside of the desired range. The results are dramatic and the effect is satisfactory for now.